+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [Unreleased]
+
+- Added `--db` option to `cldfviz.erd` command to allow bypassing SQLite creation.
+
 
 ## [v0.12.0] - 2023-02-06